CS6018_App_Sys+Design/lab08_ml_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow~=11.1.0
transformers==4.35.2
torch==2.1.1